requirements.txt
pyrofork
tgcrypto
pyleaves
python-dotenv
psutil
pillow
yt-dlp